refactor: create 'offers' and 'reviews' mock files
